requirements.txt
flask[async]
GitPython
aiofiles
sqlalchemy
flask_migrate
flask_sqlalchemy
python-dotenv
psycopg2
marshmallow
requests
